fastapi/pyproject.toml
[project]
name = "presenton-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.21.0",
    "anthropic>=0.60.0",
    "asyncpg>=0.30.0",
    "chromadb>=1.0.15",
    "dirtyjson>=1.0.8",
    "docling>=2.43.0",
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.11.0",
    "google-genai>=1.28.0",
    "nltk>=3.9.1",
    "openai>=1.98.0",
    "pathvalidate>=3.3.1",
    "python-jose[cryptography]>=3.3.0",
    "pdfplumber>=0.11.7",
    "pytest>=8.4.1",
    "python-pptx>=1.0.2",
    "redis>=6.2.0",
    "sqlmodel>=0.0.24",
    "tavily-python>=0.5.0",
    "annoy==1.17.3",
    "torch>=2.2.0",
    "transformers>=4.36.0",
    "colbert-ir>=0.2.8",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "uvicorn>=0.24.0",
    "tqdm>=4.66.1",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "requests>=2.31.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"

[tool.setuptools]
packages = ['api', 'enums', 'chroma', 'models', 'services', 'constants']

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]